Moved chart to chart_helper based on Andrew's comment. Refer to pull 439
